feat: Add SendCoinsRequestDTO and update balance type to Long in models and DTOs

+++ api/src/main/java/com/lps/api/dtos/StudentRegisterDto.java
@@ -7,7 +7,7 @@ public record StudentRegisterDto(
         String email,
         String password,
         String cpf,
-        Double balance,
+        Long balance,
         String rg,
         Long courseId,
         Address address) {
